website/requirements.txt
alembic==1.13.2
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
flask==3.0.3
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
itsdangerous==2.2.0
jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
